Use the C implementation of psycopg
The `psycopg` package defaults to a pure Python implementation.

To opt into the faster C implementation, one has to use the `c` extra.

This differs from the `psycopg2` package (which contrary to the what
the name suggests, is the older package), where the base package
included the C implementation.

See:
https://www.psycopg.org/psycopg3/docs/basic/install.html#local-installation
https://www.psycopg.org/psycopg3/docs/api/pq.html#pq-impl

GUS-W-14292989.